postgresql/Database_connect/requirements.txt
psycopg2-binary>=2.9    # PostgreSQL database adapter for Python
pandas>=1.3             # Data analysis library
sqlalchemy>=2.0.25
sqlparse=0.5.3
